refactor: deepen Kandel compliance — direction-based signaling, Area-scoped Percept, mnemonic unification — commit db90d414

+++ specification/src/main/java/xxxxx/yyyyy/zzzzz/autopoietic/autonomous/intelligence/specification/cognitive/Percept.java
@@ -7,5 +7,5 @@ public interface Percept {
 
     double intensity();
 
-    double duration();
+    long duration();
 }
